zlib_rs

Module c_api

Structs

gz_header

gzip header information passed to and from zlib routines. See RFC 1952 for more details on the meanings of these fields.

z_stream

The current stream state

Enums

internal_state

Constants

Z_ASCII

Z_BEST_COMPRESSION

Z_BEST_SPEED

Z_BINARY

Z_BLOCK

Z_BUF_ERROR

Z_DATA_ERROR

Z_DEFAULT_COMPRESSION

Z_DEFAULT_STRATEGY

Z_DEFLATED

Z_ERRNO

Z_FILTERED

Z_FINISH

Z_FIXED

Z_FULL_FLUSH

Z_HUFFMAN_ONLY

Z_MEM_ERROR

Z_NEED_DICT

Z_NO_COMPRESSION

Z_NO_FLUSH

Z_OK

Z_PARTIAL_FLUSH

Z_RLE

Z_STREAM_END

Z_STREAM_ERROR

Z_SYNC_FLUSH

Z_TEXT

Z_TREES

Z_UNKNOWN

Z_VERSION_ERROR

Type Aliases

Bytef

alloc_func

free_func

gz_headerp

in_func

out_func

uInt

uLong

uLongf

voidp

voidpc

voidpf

z_streamp