zlib_rs

Module deflate

Structs

DeflateConfig

DeflateStream

Enums

Method

Strategy

Functions

bound

For the default windowBits of 15 and memLevel of 8, this function returns a close to exact, as well as small, upper bound on the compressed size. They are coded as constants here for a reason–if the #define’s are changed, then this function needs to be changed as well. The return value for 15 and 8 only works for those exact settings.

compress

compress_bound

compress_slice

compress_slice_with_flush

compress_with_flush

copy

deflate

end

Returns

get_dictionary⚠

Safety

init

params

prime

reset

set_dictionary

set_header⚠

Safety

tune